+++ samples/Sample_29_Line.php
@@ -13,16 +13,16 @@ $section = $phpWord->addSection();
 $section->addText('Horizontal Line (Inline style):');
 $section->addLine(
     array(
-        'width' => \PhpOffice\PhpWord\Shared\Drawing::centimetersToPixels(4),
-        'height' => \PhpOffice\PhpWord\Shared\Drawing::centimetersToPixels(0),
+        'width' => \PhpOffice\PhpWord\Shared\Converter::cmToPixel(4),
+        'height' => \PhpOffice\PhpWord\Shared\Converter::cmToPixel(0),
         'positioning' => 'absolute'
     )
 );
 $section->addText('Vertical Line (Inline style):');
 $section->addLine(
     array(
-        'width' => \PhpOffice\PhpWord\Shared\Drawing::centimetersToPixels(0),
-        'height' => \PhpOffice\PhpWord\Shared\Drawing::centimetersToPixels(1),
+        'width' => \PhpOffice\PhpWord\Shared\Converter::cmToPixel(0),
+        'height' => \PhpOffice\PhpWord\Shared\Converter::cmToPixel(1),
         'positioning' => 'absolute'
     )
 );
@@ -32,13 +32,13 @@ $section->addTextBreak(1);
 $section->addText('Positioned Line (red):');
 $section->addLine(
     array(
-        'width' => \PhpOffice\PhpWord\Shared\Drawing::centimetersToPixels(4),
-        'height' => \PhpOffice\PhpWord\Shared\Drawing::centimetersToPixels(1),
+        'width' => \PhpOffice\PhpWord\Shared\Converter::cmToPixel(4),
+        'height' => \PhpOffice\PhpWord\Shared\Converter::cmToPixel(1),
         'positioning' => 'absolute',
         'posHorizontalRel' => 'page',
         'posVerticalRel' => 'page',
-        'marginLeft' => \PhpOffice\PhpWord\Shared\Drawing::centimetersToPixels(10),
-        'marginTop' => \PhpOffice\PhpWord\Shared\Drawing::centimetersToPixels(8),
+        'marginLeft' => \PhpOffice\PhpWord\Shared\Converter::cmToPixel(10),
+        'marginTop' => \PhpOffice\PhpWord\Shared\Converter::cmToPixel(8),
         'wrappingStyle' => \PhpOffice\PhpWord\Style\Image::WRAPPING_STYLE_SQUARE,
         'color' => 'red'
     )
@@ -47,8 +47,8 @@ $section->addLine(
 $section->addText('Horizontal Formatted Line');
 $section->addLine(
     array(
-        'width' => \PhpOffice\PhpWord\Shared\Drawing::centimetersToPixels(15),
-        'height' => \PhpOffice\PhpWord\Shared\Drawing::centimetersToPixels(0),
+        'width' => \PhpOffice\PhpWord\Shared\Converter::cmToPixel(15),
+        'height' => \PhpOffice\PhpWord\Shared\Converter::cmToPixel(0),
         'positioning' => 'absolute',
         'beginArrow' => \PhpOffice\PhpWord\Style\Line::ARROW_STYLE_BLOCK,
         'endArrow' => \PhpOffice\PhpWord\Style\Line::ARROW_STYLE_OVAL,