+++ tests/PhpWord/Tests/Element/TextBoxTest.php
@@ -55,11 +55,11 @@ class TextBoxTest extends \PHPUnit_Framework_TestCase
     {
         $oTextBox = new TextBox(
             array(
-                'width' => \PhpOffice\PhpWord\Shared\Drawing::centimetersToPixels(4.5),
-                'height' => \PhpOffice\PhpWord\Shared\Drawing::centimetersToPixels(17.5),
+                'width' => \PhpOffice\PhpWord\Shared\Converter::cmToPixel(4.5),
+                'height' => \PhpOffice\PhpWord\Shared\Converter::cmToPixel(17.5),
                 'positioning' => 'absolute',
-                'marginLeft' => \PhpOffice\PhpWord\Shared\Drawing::centimetersToPixels(15.4),
-                'marginTop' => \PhpOffice\PhpWord\Shared\Drawing::centimetersToPixels(9.9),
+                'marginLeft' => \PhpOffice\PhpWord\Shared\Converter::cmToPixel(15.4),
+                'marginTop' => \PhpOffice\PhpWord\Shared\Converter::cmToPixel(9.9),
                 'stroke' => 0,
                 'innerMargin' => 0,
                 'borderSize' => 1,